Made a new fragment for main activity
