+++ app/src/main/java/com/example/hsh/homesweethome/FurnitureCategoryFragment.java
@@ -4,6 +4,7 @@ import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.support.v4.app.Fragment;
+import android.support.v7.widget.GridLayoutManager;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
@@ -59,6 +60,7 @@ public class FurnitureCategoryFragment extends Fragment {
 
         locationsHorizontalRecyclerView.setAdapter(recyclerViewAdapterHorizontalLocations);
         categoryFurnitureRecyclerView.setAdapter(recyclerViewAdapterCategoryFurnitures);
+        categoryFurnitureRecyclerView.setLayoutManager(new GridLayoutManager(getActivity(),2));
 
         furnitureCategory.setText(categoryFurniture.getFurnitureCategory());
