+++ app/src/main/res/layout/content_main.xml
@@ -1,56 +1,38 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
+<RelativeLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:orientation="vertical"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
     tools:context=".MainActivity"
     tools:showIn="@layout/app_bar_main">
 
-
-    <com.arlib.floatingsearchview.FloatingSearchView
+    <android.support.v7.widget.CardView
+        android:id="@+id/search_bar"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:id="@+id/search_bar"
-        app:floatingSearch_searchBarMarginLeft="5dp"
-        app:floatingSearch_searchBarMarginRight="5dp"
-        app:floatingSearch_searchBarMarginTop="5dp"
-        app:floatingSearch_searchHint="Search"
-        app:floatingSearch_suggestionsListAnimDuration="250"
-        app:floatingSearch_showSearchKey="false"
-        app:floatingSearch_leftActionMode="showSearch"
-        app:floatingSearch_close_search_on_keyboard_dismiss="true">
+        app:cardCornerRadius="5dp"
+        android:layout_margin="5dp">
 
-        <FrameLayout
-            android:id="@+id/fragment_container"
+        <SearchView
             android:layout_width="match_parent"
+            android:queryHint="Search..."
+            android:iconifiedByDefault="false"
             android:layout_height="wrap_content">
+        </SearchView>
 
-            <ImageView
-                android:layout_width="match_parent"
-                android:layout_height="137dp"
-                android:layout_marginTop="50dp"
-                android:paddingLeft="10dp"
-                android:paddingRight="10dp"
-                android:src="@drawable/main_page_banner"
-                android:contentDescription="TODO" />
+    </android.support.v7.widget.CardView>
 
 
-            <android.support.v7.widget.RecyclerView
-                android:id="@+id/recyclerview_id"
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:padding="8dp"
-                android:layout_marginTop="170dp"
-                app:layout_behavior="@string/appbar_scrolling_view_behavior"/>
+    <FrameLayout
+        android:layout_below="@+id/search_bar"
+        android:id="@+id/fragment_container"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content">
 
-        </FrameLayout>
+    </FrameLayout>
 
-    </com.arlib.floatingsearchview.FloatingSearchView>
 
-    <!--<include layout="@layout/furniture_type_and_cards"/>-->
-
-</LinearLayout>
+</RelativeLayout>
