+++ app/src/main/res/layout/furniture_card.xml
@@ -8,7 +8,7 @@
     android:layout_width="170dp"
     android:layout_height="wrap_content"
     xmlns:cardview="http://schemas.android.com/apk/res-auto"
-    android:layout_margin="4dp"
+    android:layout_margin="10dp"
     cardview:cardCornerRadius="8dp">
 
     <LinearLayout
