+++ app/src/main/res/layout/furniture_category_fragment.xml
@@ -1,19 +1,21 @@
 <?xml version="1.0" encoding="utf-8"?>
 <RelativeLayout
-    xmlns:android="http://schemas.android.com/apk/res/android" android:layout_width="match_parent"
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
     android:layout_height="match_parent">
 
+
     <LinearLayout
-        android:weightSum="15"
+
         android:orientation="vertical"
         android:layout_width="match_parent"
         android:layout_height="match_parent">
 
         <android.support.v7.widget.RecyclerView
-            android:layout_weight="1"
             android:id="@+id/locations_horizontal_recycler"
             android:layout_width="match_parent"
             android:orientation="horizontal"
+            android:layout_margin="5dp"
             android:layout_height="wrap_content">
 
         </android.support.v7.widget.RecyclerView>
@@ -29,12 +31,14 @@
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:layout_weight="1"
+                android:paddingLeft="20sp"
+                android:textStyle="bold"
                 android:text="Category Name"/>
 
             <ImageView
                 android:layout_width="20sp"
                 android:layout_height="20sp"
-                android:layout_marginHorizontal="10dp"
+                android:layout_marginRight="20sp"
                 android:src="@drawable/filter"/>
 
 
@@ -42,13 +46,10 @@
 
         <android.support.v7.widget.RecyclerView
             android:orientation="vertical"
-            android:layout_weight="14"
             android:id="@+id/category_furnitures_recycler_view"
-            android:layout_width="match_parent"
+            android:layout_width="wrap_content"
+            android:layout_gravity="center_horizontal"
             android:layout_height="wrap_content">
-
-
-
 
         </android.support.v7.widget.RecyclerView>
