+++ app/src/main/res/layout/furniture_location_card.xml
@@ -1,20 +1,34 @@
 <?xml version="1.0" encoding="utf-8"?>
 <android.support.v7.widget.CardView
     xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_margin="4dp"
     android:layout_width="100dp"
     android:layout_height="100dp">
 
-    <TextView
-        android:id="@+id/location"
-        android:layout_gravity="center"
-        android:text="Location"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content" />
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:orientation="vertical">
+
+        <ImageView
+            android:id="@+id/location_image"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:src="@color/dark_gray"
+            android:layout_weight="1"/>
+
+        <TextView
+
+            android:textStyle="bold"
+            android:id="@+id/location"
+            android:textAlignment="center"
+            android:textSize="8dp"
+            android:text="Location"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content" />
 
 
-
-
-
+    </LinearLayout>
 
 
 </android.support.v7.widget.CardView>
